backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "sinavhazirlama-backend"
version = "1.0.0"
description = "Türk öğretmenler için sınav sorusu üretim platformu"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Sinavhazirlama Team" }]
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "alembic>=1.13.1",
    "pydantic>=2.7.1",
    "pydantic-settings>=2.2.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "google-generativeai>=0.5.4",
    "python-docx>=1.1.2",
    "google-api-python-client>=2.131.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "python-multipart>=0.0.9",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.1",
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "anyio>=4.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
